refactor(ingestor): extract shared iNat building blocks to inat-shared.ts
Lift CC_LICENSES, the tier cascade (buildTiers/Tier), the square→medium URL
substitution, and the UA/base-url constants out of client.ts into a shared
module so the upcoming top-N candidate sourcer reuses one license allowlist and
one geographic cascade. Pure refactor: fetchInatPhoto behavior and its 8 tests
are unchanged; the prior export names are re-exported for backward compat.
